% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cache_get}
\alias{cache_get}
\title{Get one element from the computation cache.}
\usage{
cache_get(name)
}
\arguments{
\item{name}{Name of cache value.}
}
\value{
The cached value, or NULL if not found. Returns invisibly if the
value was originally set with an invisible value.
}
\description{
Get one element from the computation cache.
}
