% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cache_list}
\alias{cache_list}
\title{List all names in the computation cache.}
\usage{
cache_list()
}
\value{
A character vector of cache entry names.
}
\description{
List all names in the computation cache.
}
