% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cache_set}
\alias{cache_set}
\title{Set one element of the computation cache.}
\usage{
cache_set(name, value)
}
\arguments{
\item{name}{Name of cache value.}

\item{value}{The new value.}
}
\value{
The value (preserving visibility).
}
\description{
Set one element of the computation cache.
}
