% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{\%at\%}
\alias{\%at\%}
\title{Apply a function in infix style.}
\usage{
f \%at\% x
}
\arguments{
\item{f}{A function to apply.}

\item{x}{The argument to pass to \code{f}.}
}
\value{
The result of \code{f(x)}.
}
\description{
\code{f \%at\% x} is equivalent to \code{f(x)}.
In some cases this is convenient.
Especially to prevent having to put brackets around \code{x} if \code{x} is a long piece of code.
}
