% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{hash_str}
\alias{hash_str}
\title{Turn an arbitrary string into a hash digest.}
\usage{
hash_str(x)
}
\arguments{
\item{x}{The string to hash.}
}
\value{
A hash digest string.
}
\description{
Turn an arbitrary string into a hash digest.
}
