% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{cache_rm}
\alias{cache_rm}
\title{Clears the computation cache.}
\usage{
cache_rm(names = NULL)
}
\arguments{
\item{names}{Remove only the given names.}
}
\description{
Clears all values from the cache, making it empty again.
}
